% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Onassis-methods.R
\docType{methods}
\name{annot}
\alias{annot}
\alias{annot,data.frame,character,character-method}
\alias{annot}
\title{\code{annot}}
\usage{
annot(input = NA, dictType = NA, dictionary = NA, ...)

\S4method{annot}{data.frame,character,character}(input, dictType = "OBO",
  dictionary = NA, dictoutdir = getwd(), d_synonymtype = "EXACT",
  taxID = 0, annot_out = getwd(), paramValueIndex = 0,
  searchstrategy = "CONTIGUOUS_MATCH", casematch = "CASE_INSENSITIVE",
  stemmer = "PORTER", stopwords = "NONE", orderindependentlookup = "OFF",
  findallmatches = "NO", e_synonymtype = "EXACT_ONLY",
  multipledocs = FALSE)
}
\description{
This method returns and sets the location of the dictionary.
}
\examples{

geo_chip <- readRDS(system.file('extdata', 'vignette_data',
'GEO_human_chip.rds', package='Onassis'))

obo <- system.file('extdata', 'sample.cs.obo', package='OnassisJavaLibs')
onassis_results <- annot(geo_chip, "OBO", dictionary=obo)
entities <- entities(onassis_results)
entities <- entities[sample(nrow(entities), 30),]
}
