% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllFunctions.R
\name{annotate}
\alias{annotate}
\title{\code{annotate}}
\usage{
annotate(inputFileorDf, dictionary, options = NA, outDir = tempdir(),
  multipleDocs = FALSE)
}
\arguments{
\item{inputFileorDf}{the file, directory, or data frame to annotate}

\item{dictionary}{the Conceptmapper dictionary file}

\item{options}{the options to run the finder}

\item{outDir}{the directory where to store the output files from onassis}

\item{multipleDocs}{TRUE if the input file is one but contains multiple documents}
}
\value{
dataframe of annotations
}
\description{
this function runs the entityfinder on given input
}
\examples{
obo <- system.file('extdata', 'sample.cs.obo', package='OnassisJavaLibs')
sample_dict <- CMdictionary(inputFileOrDb=obo, outputdir=getwd(), synonymType='ALL')
myopts <- new('CMoptions')
paramValueIndex(myopts) <- 40
entities <- annotate(system.file('extdata', 'vignette_data',
 'GEO_human_chip.rds', package='Onassis'), options = myopts,
 dictionary=sample_dict)

}
