% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalAnalysis.R
\name{annotateTissueDisease}
\alias{annotateTissueDisease}
\title{\code{annotateTissueDisease}}
\usage{
annotateTissueDisease(geo_metadb_path, gsm_list, tissue_obo, disease_obo,
  outdir, height_threshold, score_matrix)
}
\arguments{
\item{geo_metadb_path}{The full path of the directory where the GEOmetadb.sqlite file is stored}

\item{gsm_list}{A list of GEO sample ids (GSMs) to annotate with tissue and disease concepts}

\item{tissue_obo}{The obo ontology containing concepts to identify tissues/cell lines}

\item{disease_obo}{The obo ontology containing concepts to identify diseases}

\item{outdir}{The directory where the results will be stored}

\item{height_threshold}{The percentage of clusters to merge based on the height of the dendrogram produced by the hclust method. Height_threshold is defined in the range [0, 1].}

\item{score_matrix}{A matrix where rows represent units (GRanges or genes) and columns represent GSMs.}
}
\value{
A list of the tissue semantic sets defined by Onassis. For each tissue, a list of diseases and for each disease the columns of the score_matrix that were annotated with a given tissue and a given disease based
}
\description{
annotateTissueDisease connects to the GEOmetadb to retrieve the metadata of the samples provided in the gsm_list parameter. A dictionary for tissues/cell lines is built rom the tissue_obo file provided. All the samples' metadata are annotated with tissue concepts from the tissue_obo and samples are clustered based on the semantic similarity (default measure) of the samples. In particular, the samples are clustered and then the function cutree is used to cut the clustering tree at the specified height. Tissue semantic sets are then created associating all the samples that are similar above the provided height_threshold. Within each semantic set, samples are annotated with disease concepts from disease_obo parameter. For each disease the function retrieves the columns of the score matrix.
}
\examples{
if(!file.exists(file.path(getwd(), 'GEOmetadb.sqlite'))){
     message('To run this example please copy GEOmetadb.sqlite in your current working directory')
     } else{
     geo_metadb_path <- getwd()
     score_matrix <- readRDS(system.file('extdata', 'score_matrix.rds', package='Onassis'))
     gsm_list <- colnames(score_matrix)
     tissue_obo <- system.file('extdata', 'sample.cs.obo', package='OnassisJavaLibs')
     disease_obo <- system.file('extdata', 'sample.do.obo', package='OnassisJavaLibs')
     outdir = getwd()
     height_threshold <- 0.4
     result_list <- annotateTissueDisease(geo_metadb_path, gsm_list, tissue_obo, disease_obo, outdir, height_threshold, score_matrix)
     }
}
