% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/CMoptions-methods.R
\docType{methods}
\name{CaseMatch}
\alias{CaseMatch}
\alias{CaseMatch<-}
\alias{CaseMatch,CMoptions-method}
\alias{CaseMatch<-,CMoptions-method}
\title{\code{CaseMatch}}
\usage{
CaseMatch(x)

CaseMatch(x) <- value

\S4method{CaseMatch}{CMoptions}(x)

\S4method{CaseMatch}{CMoptions}(x) <- value
}
\arguments{
\item{x}{instance of class \code{\link{CMoptions}}}

\item{value}{One of the following :
\itemize{
\item{CASE_IGNORE}{ Fold everything to lowercase for matching}
\item{CASE_INSENSITIVE}{ Fold only tokens with initial caps to lowercase}
\item{CASE_FOLD_DIGITS}{ Fold all (and only) tokens with a digit}
\item{CASE_SENSITIVE}{ Perform no case folding}
}}
}
\value{
The CaseMatch corresponding to the current options when used as get, the new CMoptions object with updated parameters when used to set.
}
\description{
Method to get and set the CaseMatch parameter of \code{\link{CMoptions}} instances

\code{CaseMatch<-}
}
\examples{
opts <- CMoptions()
CaseMatch(opts)
opts <- CMoptions()
CaseMatch(opts) <- 'CASE_SENSITIVE'
}
