% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{CMdictionary-class}
\alias{CMdictionary-class}
\title{Class that stores a Conceptmapper dictionary}
\description{
CMdictionary is a class that wraps a Conceptmapper ccp-nlp Java dictionary. Its methods allow the creation of a dictionary from OBO ontologies in OBO or OWL format. Different options to build the dictionary are available.
}
\details{
The following methods can be applied to CMdictionary instances

\code{\link{dict_location}} \cr
\code{\link{dict_location<-}} \cr
\code{\link{dictInfo}} \cr
\code{\link{dictInfo<-}} \cr
\code{\link{dictRef}} \cr
\code{\link{dictRef<-}} \cr

To show the available dictionary types use the function
\code{\link{dictTypes}} \cr
}
\section{Slots}{

\describe{
\item{\code{dict_location}}{The path of the created dictionary file}

\item{\code{dictRef}}{Reference to the java object representing the dictionary}

\item{\code{dictInfo}}{Information about how dictionary has been created. It is a list with the following fields
\itemize{
\item{Dictionary Type:{The type of dictionary}
 \itemize{
      \item{OBO}{A dictionary that has been created by An OBO file}
      \item{ENTREZ}{Entrez genes dictionary}
      \item{TARGET}{Entrez genes dictionary, Histone marks and Histone modifications}
      \item{CMDICT}{A previously created dictionary file in the Conceptmapper XML format}
 }
}

\item{SynonymType: {The type of synonyms to consider when building the dictionary for Conceptmapper. For further detail \url{http://owlcollab.github.io/oboformat/doc/obo-syntax.html}}
\itemize{
\item{EXACT}{}
\item{BROAD}{}
\item{NARROW}{}
\item{RELATED}{}
\item{ALL}{}
}
}

\item{dictSource: {The OBO/OWL dictionary file to convert to a Conceptmapper dictionary in case the type is OBO. The XML file in case the type is CMDICT}}

\item{taxID The NCBI taxon identifier for species to create the Entrez gene dictionary (e.g 9606 for Mus musculus)}
}}
}}

\examples{
dict <- new('CMdictionary')
}
