% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllConstructors.R
\name{CMdictionary}
\alias{CMdictionary}
\title{\code{CMdictionary}}
\usage{
CMdictionary(inputFileOrDb = NULL, dictType = "OBO",
  outputDir = tempdir(), synonymType = "EXACT", taxID = 0,
  outputDirOp = TRUE)
}
\arguments{
\item{inputFileOrDb}{The local OBO/OWL ontology to be converted into an XML Conceptmapper dictionary or the URL of a OBO/OWL file. If inputFileOrdb is NA and the \code{dicType} parameter is not the generic OBO then the method tries to download the corresponding dictionary from the available repositories. For ENTREZ and TARGET dictionary types a file named gene_info.gz will be automatically downloaded from \url{ftp://ncbi.nlm.nih.gov/gene/data/gene_info.gz} if a valid path is not provided by the user. Alternatively  the name of an annotation package of the type Org.xx.eg.db from Bioconductor can be used. In this case the gene unique identifiers and their alternative identifiers will be retrieved from the annotation database without the need of downloading a gene_info file.}

\item{dictType}{the type of input dictionary
\describe{
\item{OBO}{A dictionary that has been created by An OBO file}
\item{ENTREZ}{Entrez genes dictionary}
\item{TARGET}{Entrez genes dictionary, Histone marks and Histone modifications}
\item{CMDICT}{A previously created dictionary file in the Conceptmapper XML format}
}}

\item{outputDir}{the directory where the XML conceptmapper dictionary will be stored. Defaults to the tmp system's directory}

\item{synonymType}{The type of synonyms to consider when building the dictionary for Conceptmapper. For further detail \url{http://owlcollab.github.io/oboformat/doc/obo-syntax.html}. Default: EXACT
\describe{
\item{EXACT}{}
\item{ALL}{}
}}

\item{taxID}{the taxonomy identifier of the organism when the dictionary type is ENTREZ or TARGET. If 0 all the taxonomies will be included in the new dictionary.}

\item{outputDirOp}{set to TRUE to clean the directory before creating the dictionary}
}
\value{
An object of type \code{\link{CMdictionary-class}} that can be used to annotate text with the \code{\link{EntityFinder}}.
}
\description{
Constructor method for creating instances of class \code{\link{CMdictionary-class}}. The created Conceptmapper dictionary will be stored as an XML file in the file system.
}
\examples{
\dontrun{
#' ##This might take some time to download the dictionary
dict <- CMdictionary(dictType = 'TARGET', inputFileOrDb='org.Hs.eg.db')

dict_file <- system.file('extdata', 'sample.cs.obo', package='OnassisJavaLibs')
dictionary <- CMdictionary(dictType='OBO', inputFileOrDb=dict_file)
}
}
