% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllConstructors.R
\name{CMoptions}
\alias{CMoptions}
\title{\code{CMoptions}}
\usage{
CMoptions(SearchStrategy = "CONTIGUOUS_MATCH",
  CaseMatch = "CASE_INSENSITIVE", Stemmer = "NONE", StopWords = "NONE",
  OrderIndependentLookup = "ON", FindAllMatches = "YES",
  SynonymType = "ALL", paramValueIndex = NA)
}
\arguments{
\item{SearchStrategy}{The matching strategy for finding concepts in the input text
\itemize{
\item{CONTIGUOUS_MATCH}{Longets match of contiguous tokens within enclosing span}
\item{SKIP_ANY_MATCH}{Longest match of not-necessarily contiguous tokens}
\item{SKIP_ANY_MATCH_ALLOW_OVERLAP}{Longest match of not-necessarily contiguous tokens, overlapping matches are allowed}
}}

\item{CaseMatch}{\itemize{
\item{CASE_IGNORE}{Fold everything to lowercase for matching}
\item{CASE_INSENSITIVE}{Fold only tokens with initial caps to lowercase}
\item{CASE_FOLD_DIGITS}{Fold all (and only) tokens with a digit}
\item{CASE_SENSITIVE}{Perform no case folding}
}}

\item{Stemmer}{\itemize{
\item BIOLEMMATIZER {A stemmer specific for biomedical literature}
\item PORTER {A stemmer that removes the commoner morphological and inflexional endings from words in English}
\item NONE {No word stemming}
}}

\item{StopWords}{\itemize{
\item PUBMED {A list of stop words obtained analyzing Pubmed papers}
\item NONE {No stop words }
}}

\item{OrderIndependentLookup}{\itemize{
\item ON {Ordering within span is ignored (i.e. 'Breast cancer' would equal 'Cancer breast') }
\item OFF {Ordering is taken into consideration}
}}

\item{FindAllMatches}{\itemize{
\item YES {All the matches within the span are found }
\item NO {Only the longest match within the span will be returned}
}}

\item{SynonymType}{\itemize{
\item EXACT_ONLY {Only exact synonyms are considered }
\item ALL {All synonym types are included}
}}

\item{paramValueIndex}{An integer value to index the 576 parameter combinations}
}
\value{
instance of the class CMoptions set to the default combination of parameters
}
\description{
This constructor creates an object of type CMoptions
}
\examples{
op <- CMoptions()
}
