% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEOHandler.R
\name{connectToGEODB}
\alias{connectToGEODB}
\title{\code{connectToGEODB}}
\usage{
connectToGEODB(sqliteFileName = NULL, download = FALSE)
}
\arguments{
\item{sqliteFileName}{optional SQLite file path of the SQLite database if already downloaded}

\item{download}{If TRUE allow the automatic downloading of the database file.}
}
\value{
A connection to the GEOmetadb
}
\description{
This method allows users to connect to the GEOmetadb downloaded. If no parameter is provided than the function retrieves the database in sqlite format and returns a connection to query the database
}
\examples{
\dontrun{
 geo_connection <- connectToGEODB(download=TRUE)
 }
 geo_con <- connectToGEODB()
}
