% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRAdbHandler.R
\name{connectToSRADB}
\alias{connectToSRADB}
\title{\code{connectToSRADB}}
\usage{
connectToSRADB(sqliteFileName = NULL)
}
\arguments{
\item{sqliteFileName}{optional SQLite file path of the SQLite database if already downloadd}
}
\value{
A connection to the SRAdb
}
\description{
This method allows users to connect to the SRAdb downloaded. If no parameter is provided than the function retrieves the database in sqlite format and return a connection to query the database
}
\examples{
if(file.exists('SRAdb.sqlite')){
 sra_con <- connectToSRADB('SRAdb.sqlite')
}else{
 print('You need to download SRAdb.sqlite to run this example')
}
}
