% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Onassis-methods.R
\docType{methods}
\name{dict}
\alias{dict}
\alias{dict<-}
\alias{dict<-,Onassis-method}
\alias{dict<-}
\alias{dict,Onassis-method}
\alias{dict}
\title{Method dictionary}
\usage{
dict(object)

dict(object) <- value

\S4method{dict}{Onassis}(object) <- value

\S4method{dict}{Onassis}(object)
}
\value{
The Onassis object

Ontology object
}
\description{
Method dictionary

Method dict<-

This method sets the dictionary of Onassis

This method shows the ontology.
}
\examples{
onassis <- Onassis()
dict(onassis) <- file.path(getwd(), 'obo.obo')

o <- Onassis()
dict(o)

}
