% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/CMdict-methods.R
\docType{methods}
\name{dict_location}
\alias{dict_location}
\alias{dict_location<-}
\alias{dict_location,CMdictionary-method}
\alias{dict_location<-,CMdictionary-method}
\title{\code{dict_location}}
\usage{
dict_location(.Object)

dict_location(.Object) <- value

\S4method{dict_location}{CMdictionary}(.Object)

\S4method{dict_location}{CMdictionary}(.Object) <- value
}
\arguments{
\item{.Object}{An instance of class CMdictionary}

\item{value}{the path of the new dictionary}
}
\value{
the path of the dictionary in the file system
}
\description{
A method to get and set the location of  \code{\link{CMdictionary}}

\code{dict_location<-}
}
\examples{
dictionary <- new('CMdictionary')
loc <- dict_location(dictionary)
dictionary <- new('CMdictionary')
dict_location(dictionary) <- getwd()
}
