% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/CMdict-methods.R
\docType{methods}
\name{dictInfo}
\alias{dictInfo}
\alias{dictInfo<-}
\alias{dictInfo,CMdictionary-method}
\alias{dictInfo<-,CMdictionary-method}
\title{Get and Set the dicInfo slot for the \code{\link{CMdictionary}} class}
\usage{
dictInfo(.Object)

dictInfo(.Object) <- value

\S4method{dictInfo}{CMdictionary}(.Object)

\S4method{dictInfo}{CMdictionary}(.Object) <- value
}
\arguments{
\item{.Object}{An instance of class \code{\link{CMdictionary}}}

\item{value}{list of details about the dictionary}
}
\value{
list of details about the dictionary

object of class \code{\link{CMdictionary}}
}
\description{
Get and Set the dicInfo slot for the \code{\link{CMdictionary}} class

Method to get and set a list of info about the dictionary

\code{dictInfo<-}
}
\examples{
dictionary <- new('CMdictionary')
dictInfo(dictionary)
dictionary <- new('CMdictionary')
dictInfo(dictionary) <-
list(Dictionary_type =  'ENTREZ from OrgDb', Dictionary_source ='OrgDb')
}
