% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllFunctions.R
\name{dictionaryInfo}
\alias{dictionaryInfo}
\title{\code{dictionaryInfo}}
\usage{
dictionaryInfo(CMdictionary)
}
\arguments{
\item{CMdictionary}{CMdictionary object of the conceptmapper dictionary}
}
\value{
list of details about the dictionary
}
\description{
This function shows the list of details of the conceptmapper dictionary
}
\examples{
obo <- system.file('extdata', 'sample.cs.obo', package='OnassisJavaLibs')
sample_dict <- CMdictionary(inputFileOrDb=obo, outputdir=getwd(), synonymType='ALL')
dictionaryInfo(sample_dict)
}
