% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllFunctions.R
\name{dictionaryTypes}
\alias{dictionaryTypes}
\title{\code{dictTypes}}
\usage{
dictionaryTypes()
}
\value{
the list of dictionary types available
}
\description{
This function shows a lis of the pre-defined conceptmapper dictionary types
}
\examples{
types <- dictionaryTypes()
}
