% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMdict-methods.R
\name{dictTypes}
\alias{dictTypes}
\title{\code{dictTypes}}
\usage{
dictTypes()
}
\value{
list of types that can be used to create a Conceptmapper dictionary
}
\description{
A function to show the pre-defined Conceptmapper dictionary types wwith their descriptions
}
\examples{
dictTypes()
}
