% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Onassis-methods.R
\docType{methods}
\name{entities}
\alias{entities}
\alias{entities<-}
\alias{entities<-,Onassis-method}
\alias{entities<-}
\alias{entities,Onassis-method}
\alias{entities}
\title{Method entities}
\usage{
entities(object)

entities(object) <- value

\S4method{entities}{Onassis}(object) <- value

\S4method{entities}{Onassis}(object)
}
\value{
The Onassis object

Ontology object
}
\description{
Method entities

Method entities<-

This method sets the entities of Onassis

This method shows the entities matrix in Onassis.
}
\examples{
onassis <- Onassis()
entities(onassis) <- matrix()

o <- Onassis()
entities(o)

}
