% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{EntityFinder-class}
\alias{EntityFinder-class}
\title{EntityFinder class to create a Conceptmapper instance}
\description{
EntityFinder is a class that wraps a Conceptmapper pipeline using the CCP UIMA Type System \url{https://github.com/UCDenver-ccp/ccp-nlp}. The pipeline includes a sentence detector, offset tokenizer and retrieves concepts from dictionaries built from  OBO/OWL formatted ontology files.
}
\details{
The following methods can be applied to EntityFinder

\code{\link{findEntities}} \cr
\code{\link{annotateDF}} \cr
The methods can be automatically called using the function
\code{\link{EntityFinder}} \cr
}
\section{Slots}{

\describe{
\item{\code{typeSystemRef}}{The reference to the Java object representing the type system}
}}

\examples{
finder <- new('EntityFinder')
}
