% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEOHandler.R
\name{experiment_types}
\alias{experiment_types}
\title{\code{experiment_types}}
\usage{
experiment_types(GEOcon)
}
\arguments{
\item{GEOcon}{connection to the SQLite GEOmetadb databse}
}
\value{
A character vector with all the possible experiment values
}
\description{
This method retrierves the experiment types stored in GEOmetadb
}
\examples{
if(file.exists('GEOmetadb.sqlite')){
 geo_con <- connectToGEODB('GEOmetadb.sqlite')
 experiments <- experiment_types(geo_con)
}else{
 print('You need to download GEOmetadb.sqlite to run this example')
}
}
