% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllFunctions.R
\name{filterTerms}
\alias{filterTerms}
\title{\code{filterTerms}}
\usage{
filterTerms(annotated_df, termlist = c())
}
\arguments{
\item{annotated_df}{a data frame resulting from the annotation process of Onassis}

\item{termlist}{the list of terms to be filtered out from the annotations}
}
\value{
a data frame with removed annotations
}
\description{
filterTerms allows users to remove a set of defined terms from the results of the annotation provided by Onassis
}
\examples{
metadatafile <- readRDS(system.file('extdata', 'vignette_data',
'GEO_human_chip.rds', package='Onassis'))
healthy_gsms <- findHealthy(metadatafile)
}
