% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMoptions-methods.R
\docType{methods}
\name{FindAllMatches,CMoptions-method}
\alias{FindAllMatches,CMoptions-method}
\title{\code{FindAllMatches}}
\usage{
\S4method{FindAllMatches}{CMoptions}(x)
}
\arguments{
\item{x}{instance of class \code{\link{CMoptions}}}
}
\value{
The FindAllMatches corresponding to the current options when used as get, the new CMoptions object with updated parameters when used to set.
}
\description{
Method to get and set the FindAllMatches parameter
}
\examples{
opts <- CMoptions()
FindAllMatches(opts)

}
