% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/CMoptions-methods.R
\docType{methods}
\name{FindAllMatches}
\alias{FindAllMatches}
\alias{FindAllMatches<-}
\alias{FindAllMatches<-,CMoptions-method}
\title{Method FindAllMatches}
\usage{
FindAllMatches(x)

FindAllMatches(x) <- value

\S4method{FindAllMatches}{CMoptions}(x) <- value
}
\arguments{
\item{value}{\itemize{
\item YES {All the matches within the span are found }
\item NO {Only the longest match within the span will be returned}
}}
}
\description{
Method FindAllMatches

Method FindAllMatches<-

\code{FindAllMatches<-}
}
\examples{
opts <- CMoptions()
FindAllMatches(opts) <- 'YES'
}
