% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/EntityFinder-methods.R
\docType{methods}
\name{findEntities}
\alias{findEntities}
\alias{findEntities,EntityFinder,character,logical,character,CMoptions,CMdictionary-method}
\title{Method findEntities}
\usage{
findEntities(object, inputDirOrFile, multipleDocs = FALSE,
  outDir = tempdir(), configOpt, cmDict)


  \S4method{findEntities}{EntityFinder,character,logical,character,CMoptions,CMdictionary}(object,
  inputDirOrFile, multipleDocs = FALSE, outDir = tempdir(), configOpt,
  cmDict)
}
\arguments{
\item{object}{instance of the class EntityFinder}

\item{inputDirOrFile}{the directory where the files to annotate are stored or the text file to annotate. A single file containing in each row sample names, the | symbol and the description of the sample is also allowed.}

\item{multipleDocs}{TRUE if a single file containing different text sources has been given as inputDirOrFile. FALSE if each text is in a separate file. Defaults to FALSE}

\item{outDir}{The directory where the Conceptmapper annotated files are stored. Default: the system tmp directory.}

\item{configOpt}{Object of type CMoptions in which the parameters to run Conceptmapper are stored}

\item{cmDict}{Instance of class \code{\link{CMdictionary}} or the file path of an already created CMdictionary}
}
\value{
A data frame of annotations containing the sample name, the id of the OBO concept, the corresponding name, the part of the text containing the annotation
}
\description{
Method findEntities

This method finds concepts from a Conceptmapper Dictionary of type \code{\link{CMdictionary}} in a given directory or in a single pipe separated file containing a named document in each row, with a specified configuration of type \code{\link{CMoptions}}. This is a method of the \code{\link{EntityFinder-class}}
}
\examples{
obo <- system.file('extdata', 'sample.cs.obo', package='OnassisJavaLibs')
dict <- CMdictionary(inputFileOrDb=obo, outputDir=getwd(), synonymType='ALL')

opts <- CMoptions()
ef <- new('EntityFinder')
annotations <- findEntities(ef,
system.file('extdata', 'test_samples', 'test_samples.txt', package='Onassis'), multipleDocs=TRUE, outDir=getwd(),
 configOpt=opts, cmDict=dict)

}
