% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllFunctions.R
\name{findHealthy}
\alias{findHealthy}
\title{\code{findHealthy}}
\usage{
findHealthy(metadata_df)
}
\arguments{
\item{metadata_df}{a data frame where the first column corresponds to the identifier of a sample #' and the other columns the metadata relative to the sample}
}
\value{
a data frame with healthy samples annotated as 'Healthy'
}
\description{
findHealthy annotates as 'Healthy' the samples whose metadata matches with one of the elements of a list of sentence used to describe the normal disease state or healthy.
}
\examples{
metadatafile <- readRDS(system.file('extdata', 'vignette_data',
'GEO_human_chip.rds', package='Onassis'))

healthy_gsms <- findHealthy(metadatafile)
}
