% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEOHandler.R
\name{connectToGEODB}
\alias{connectToGEODB}
\alias{connectToGEODB,}
\alias{GEOHandler-function}
\title{\code{connectToGEODB}}
\usage{
connectToGEODB(sqliteFilePath = NULL, download = FALSE, destdir = getwd())
}
\arguments{
\item{sqliteFilePath}{optional SQLite full file path of the SQLite database if already downloaded}

\item{download}{If TRUE allow the automatic downloading of the database file.}

\item{destdir}{optional destination directory. Current working directory is the default}
}
\value{
A connection to the GEOmetadb
}
\description{
This method allows users to connect to the GEOmetadb downloaded. If no parameter is provided than the function retrieves the database in sqlite format and returns a connection to query the database
}
\examples{
\dontrun{
 geo_connection <- connectToGEODB(download=TRUE)
 }
 if(file.exists('GEOmetadb.sqlite')){
    geo_con <- connectToGEODB()
 } else {
 message('Please provide GEOmetadb.sqlite file')
 }
}
