% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEOHandler.R
\name{getGEOMetadata}
\alias{getGEOMetadata}
\title{\code{getGEOMetadata}}
\usage{
getGEOMetadata(geo_con, experiment_type = NA, organism = NA,
  gpl = NA)
}
\arguments{
\item{geo_con}{connection to the SQLite GEOmetadb databse}

\item{experiment_type}{The type of experiment. Allowed values can be obtained through the function \code{experiment_types}}

\item{organism}{Optional type of organism. Allowed species can be obtained using the function \code{organism_types}. If no organism is passed as parameter the query will retrieve all the organisms}

\item{gpl}{Optional platform identifier in case a platform based query has to be executed}
}
\value{
A data frame with the queried samples' metadata
}
\description{
This method retrierves the descriptive fields of the samples in GEO for a given experiment_type, organism or platform.
}
\examples{
if(file.exists('GEOmetadb.sqlite')){
 geo_con <- connectToGEODB('GEOmetadb.sqlite')
methilation <- getGEOMetadata(geo_con,
'Methylation profiling by high throughput sequencing', 'Homo sapiens')
expression <- getGEOMetadata(geo_con,
 'Expression profiling by array', 'Homo sapiens', 'GPL570')
}else{
 print('You need to download GEOmetadb.sqlite to run this example')
}
}
