% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRAdbHandler.R
\name{getSRAMetadata}
\alias{getSRAMetadata}
\title{\code{getSRAMetadata}}
\usage{
getSRAMetadata(sra_con, library_strategy, library_source = NA,
  taxon_id = NA, center_name = NA)
}
\arguments{
\item{sra_con}{connection to the SQLite SRAdb databse}

\item{library_strategy}{corresponding to the  type of experiment. Allowed values can be obtained through the function \code{library_strategies}}

\item{library_source}{corresponding to the material from the sample (GENOMIC, METAGENOMIC, TRANSCRIPTOMIC...)}

\item{taxon_id}{Optional taxon id of the organism. If no taxon_id is passed as parameter the query will retrieve all the organisms}

\item{center_name}{Optional center name identifier. Possible values can be obtained querying the database e.g. dbGetQuery(sra_con, "select distinct center_name from sample")}
}
\value{
A data frame with the queried samples' metadata
}
\description{
This method retrierves the descriptive fields of experiments and samples in SRAdb for a given library_strategy (experiment type), of a given organism and center name.
}
\examples{
if(file.exists('SRAdb.sqlite')){
   sra_con <- connectToSRADB('SRAdb.sqlite')
   geo_human_chip <- getSRAMetadata(sra_con, library_strategy='ChIP-Seq',
   library_source='GENOMIC', taxon_id=9606, center_name='GEO')
   bisulfite_seq <- getSRAMetadata(sra_con, library_strategy='Bisulfite-Seq',
    library_source='GENOMIC', taxon_id=9606, center_name='GEO')
 } else{
   print('You need to download the SRAdb.sqlite to run this example')
   }
}
