% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/similarity-methods.R
\docType{methods}
\name{groupsim}
\alias{groupsim}
\alias{groupsim,Similarity,character,character-method}
\alias{groupsim-method}
\title{Method groupsim}
\usage{
groupsim(object, termList1, termList2)

\S4method{groupsim}{Similarity,character,character}(object, termList1,
  termList2)
}
\arguments{
\item{object}{Instance of class similarity}

\item{termList1}{A vector of URIs of ontology terms in the format \url{http://purl.obolibrary.org/obo/Ontology_id} (e.g \url{http://purl.obolibrary.org/obo/BTO_0004732})}

\item{termList2}{A vector of URIs of ontology terms}
}
\value{
the semantic similarity of the two provided groups of concepts
}
\description{
Method groupsim

This method computes the semantic similarity between two groups of terms of a given ontology.
}
\examples{

sim <- new('Similarity')

dict_file <-  file.path(getwd(), 'BrendaTissue.obo')
 if(!file.exists(dict_file))
   download.file('https://sourceforge.net/projects/onassis/files/BrendaTissue.obo',
    destfile=file.path(getwd(), 'BrendaTissue.obo'))
ontology(sim) <- dict_file
pairwiseConfig(sim) <- showOpts(sim)$pairwiseMeasures[9]
groupwiseConfig(sim) <- showOpts(sim)$groupwiseMeasures[3]
similarity <- groupsim(sim, c('http://purl.obolibrary.org/obo/BTO_0004732',
'http://purl.obolibrary.org/obo/BTO_0001718'),
c('http://purl.obolibrary.org/obo/BTO_0000815', 'http://purl.obolibrary.org/obo/BTO_0001912'))
similarity

}
