% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/similarity-methods.R
\docType{methods}
\name{groupwiseConfigRef}
\alias{groupwiseConfigRef}
\alias{groupwiseConfigRef<-}
\alias{groupwiseConfigRef,Similarity-method}
\alias{groupwiseConfigRef<-,Similarity-method}
\title{\code{groupwiseConfigRef}}
\usage{
groupwiseConfigRef(object)

groupwiseConfigRef(object) <- value

\S4method{groupwiseConfigRef}{Similarity}(object)

\S4method{groupwiseConfigRef}{Similarity}(object) <- value
}
\arguments{
\item{object}{instance of class \code{\link{Similarity-class}}}

\item{value}{one of the groupwise measures. See \code{\link{groupConfig}}}
}
\value{
groupwise configured measure for the similarity object provided as input

instance of the Similarity class with the new grouprwise option.
}
\description{
This method shows the value of the groupwise configuration used to compute semantic similarities between groups of concepts.

Sets the groupwise measure to the reference of a groupwise measure to the semantic similarity between groups of concepts. For available meausres see the groupConfig function's details.
}
\examples{
sim <- new('Similarity')
groupwiseConfigRef(sim)
sim <- new('Similarity')
obo <- system.file('extdata', 'sample.cs.obo', package='OnassisJavaLibs')
groupwiseConfigRef(sim) <- 'ui'
}
