% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/similarity-methods.R
\docType{methods}
\name{icConfig}
\alias{icConfig}
\alias{icConfig<-}
\alias{icConfig,Similarity-method}
\alias{icConfig<-,Similarity-method}
\title{\code{icConfig}}
\usage{
icConfig(object)

icConfig(object) <- value

\S4method{icConfig}{Similarity}(object)

\S4method{icConfig}{Similarity}(object) <- value
}
\arguments{
\item{object}{instance of class \code{\link{Similarity-class}}}

\item{value}{the information content measure selected.}
}
\value{
The measure used to compute concepts' information content

The similarity object with the new information conten measure set
}
\description{
This method retrieves the configuration of the intrinsic information content measure

This method sets the configuration of the intrinsic information content measure by taking as parameter the short flag associated to the information measure. To have details about the available short flags see the \code{\link{pairwiseConfig}} help
}
\examples{
sim <- new('Similarity')
icConfig(sim)
sim <- new('Similarity')
icConfig(sim) <- 'sanchez'
}
