% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRAdbHandler.R
\name{library_sources}
\alias{library_sources}
\title{\code{library_sources}}
\usage{
library_sources(SRAcon)
}
\arguments{
\item{SRAcon}{connection to the SQLite SRAdb databse}
}
\value{
A character vector with all the possible library sources
}
\description{
This method retrierves the library source types stored in SRAdb
}
\examples{
if(file.exists('SRAdb.sqlite')){
   sra_con <- connectToSRADB('SRAdb.sqlite')
   sources <- library_sources(sra_con)
   head(sources)
   }else{
   print('You need to download the SRAdb.sqlite to run this example')
   }
}
