% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRAdbHandler.R
\name{library_strategies}
\alias{library_strategies}
\title{\code{library_strategies}}
\usage{
library_strategies(SRAcon)
}
\arguments{
\item{SRAcon}{connection to the SQLite SRAdb databse}
}
\value{
A character vector with all the possible library strategy values
}
\description{
This method retrierves the library strategy types stored in SRAdb
}
\examples{
if(file.exists('SRAdb.sqlite')){
   sra_con <- connectToSRADB('SRAdb.sqlite')
   strategies <- library_strategies(sra_con)
   head(strategies)
 } else{
   print('You need to download the SRAdb.sqlite to run this example')
 }
}
