% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity-methods.R
\name{listSimilarities}
\alias{listSimilarities}
\title{\code{listSimilarities}}
\usage{
listSimilarities()
}
\value{
the list of pairwise, information content and groupwise measures to compute the semantic similarities
}
\description{
This method shows a list of the possible measures to compute pairwise and groupwise semantic similarity between concepts
}
\examples{
s <- listSimilarities()
}
