% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/similarity-methods.R
\docType{methods}
\name{ontology}
\alias{ontology}
\alias{ontology<-}
\alias{ontology<-,Similarity-method}
\alias{ontology,Similarity-method}
\title{\code{ontology<-}}
\usage{
ontology(object)

ontology(object) <- value

\S4method{ontology}{Similarity}(object) <- value

\S4method{ontology}{Similarity}(object)
}
\arguments{
\item{object}{instance of class \code{\link{Similarity-class}}}

\item{value}{The path of an ontology file}
}
\value{
The Similarity object where 'ontology' slot refers to the Java graph created

Ontology object
}
\description{
This method creates a semantic graph to compute semantic similarity between concepts. It takes as input an OBO ontology in RDF, OWL or OBO format.

This method shows the ontology.
}
\examples{
sim <- new('Similarity')
obo <- system.file('extdata', 'sample.cs.obo', package='OnassisJavaLibs')
ontology(sim) <- obo
sim <- new('Similarity')
ontology(sim)
}
