% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMoptions-methods.R
\docType{methods}
\name{OrderIndependentLookup,CMoptions-method}
\alias{OrderIndependentLookup,CMoptions-method}
\title{\code{OrderIndependentLookup}}
\usage{
\S4method{OrderIndependentLookup}{CMoptions}(x)
}
\arguments{
\item{x}{instance of class \code{\link{CMoptions}}}
}
\value{
The OrderIndependentLookup corresponding to the current options when used as get, the new CMoptions object with updated parameters when used to set.
}
\description{
Method to get and set the OrderIndependentLookup parameter
}
\examples{
opts <- CMoptions()
OrderIndependentLookup(opts)

}
