% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEOHandler.R
\name{organism_types}
\alias{organism_types}
\title{\code{organism_types}}
\usage{
organism_types(geo_con)
}
\arguments{
\item{geo_con}{connection to the SQLite GEOmetadb databse}
}
\value{
A character vector with all the possible organism values
}
\description{
This method retrierves the allowed organisms in GEOmetadb
}
\examples{
if(file.exists('GEOmetadb.sqlite')){
 geo_con <- connectToGEODB('GEOmetadb.sqlite')
 species <- organism_types(geo_con)
}else{
 print('You need to download GEOmetadb.sqlite to run this example')
}
}
