% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/similarity-methods.R
\docType{methods}
\name{pairwiseConfigRef}
\alias{pairwiseConfigRef}
\alias{pairwiseConfigRef<-}
\alias{pairwiseConfigRef,Similarity-method}
\alias{pairwiseConfigRef<-,Similarity-method}
\title{\code{pairwiseConfigRef}}
\usage{
pairwiseConfigRef(object)

pairwiseConfigRef(object) <- value

\S4method{pairwiseConfigRef}{Similarity}(object)

\S4method{pairwiseConfigRef}{Similarity}(object) <- value
}
\arguments{
\item{object}{instance of class \code{\link{Similarity-class}}}

\item{value}{See details}
}
\value{
The reference to the pairwise configuration used to compute semantic similarity
}
\description{
This method retrieves the reference to the Java configuration used to compute semantic similarities.

configures the pairwise java object to compute semantic similarity between two concepts of a given ontology, by passing as input the java reference to one of the allowed pairwise semantic similarity measures. For a complete list check the details section of the function pairwiseConfig.
}
\examples{
sim <- new('Similarity')
pairwiseConfigRef(sim)
sim <- new('Similarity')
pairwiseConfigRef(sim) <- c('resnik')
}
