% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMoptions-methods.R
\docType{methods}
\name{paramValueIndex,CMoptions-method}
\alias{paramValueIndex,CMoptions-method}
\title{\code{paramValueIndex}}
\usage{
\S4method{paramValueIndex}{CMoptions}(x)
}
\arguments{
\item{x}{instance of class \code{\link{CMoptions}}}
}
\value{
The paramValueIndex corresponding to the current options when used as get, the new CMoptions object with updated parameters when used to set.
}
\description{
Method to get and set the parameter combination index corresponding to a given parameter combination. The value of the paramValueIndex lays in the range [0:575]
}
\examples{
opts <- CMoptions()
paramValueIndex(opts)

}
