% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/CMoptions-methods.R
\docType{methods}
\name{paramValueIndex}
\alias{paramValueIndex}
\alias{paramValueIndex<-}
\alias{paramValueIndex<-,CMoptions-method}
\title{Method paramValueIndex}
\usage{
paramValueIndex(x)

paramValueIndex(x) <- value

\S4method{paramValueIndex}{CMoptions}(x) <- value
}
\arguments{
\item{value}{Index corresponding to a given parameter combination}
}
\description{
Method paramValueIndex

Method paramValueIndex<-

\code{paramValueIndex<-}
}
\examples{
opts <- CMoptions()
paramValueIndex(opts) <- 2
}
