% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/similarity-methods.R
\docType{methods}
\name{samplesim}
\alias{samplesim}
\alias{samplesim,Similarity,character,character,data.frame-method}
\title{Method samplesim}
\usage{
samplesim(object, sample1, sample2, annotated_df)

\S4method{samplesim}{Similarity,character,character,data.frame}(object, sample1,
  sample2, annotated_df)
}
\arguments{
\item{object}{Instance of the class \code{\link{Similarity-class}}}

\item{sample1}{A sample ID with its annotations available in a data frame}

\item{sample2}{A sample ID with its annotations available in a data frame}

\item{annotated_df}{data frame with annotations obtained using entityFinder. The data frame should have at least a column named 'sample_id' with the sample identifier and a column named 'term_url' with the URL of the ontology terms annotating the sample. The ontology terms must belong to the ontology loaded in the Similarity class.}
}
\value{
The semantic similarity between the samples sample1 and sample2
}
\description{
Method samplesim

This method computes the semantic similarity between two named samples annotated with a group of ontolgy terms belonging to the same ontology
}
\examples{
ef <- new('EntityFinder')
opts <- new('CMoptions')
dict_file <-  file.path(getwd(), 'cmDict-BrendaTissue.xml')
 if(!file.exists(dict_file))
   download.file('https://sourceforge.net/projects/onassis/files/cmDict-BrendaTissue.xml',
    destfile=file.path(getwd(), 'cmDict-BrendaTissue.xml'))
brenda <- findEntities(ef, system.file('extdata', 'testsamples',
package='Onassis'), multipleDocs=FALSE, configOpt=opts,
 cmDict=file.path(getwd(), 'cmDict-BrendaTissue.xml'))
sim <- new('Similarity')
dict_file <-  file.path(getwd(), 'BrendaTissue.obo')
 if(!file.exists(dict_file))
   download.file('https://sourceforge.net/projects/onassis/files/BrendaTissue.obo',
    destfile=file.path(getwd(), 'BrendaTissue.obo'))
ontology(sim) <- file.path(getwd(), 'BrendaTissue.obo')
pairwiseConfig(sim) <- showOpts(sim)$pairwiseMeasures[9]
groupwiseConfig(sim) <- showOpts(sim)$groupwiseMeasures[3]
similarity <- samplesim(sim, brenda[1,1], brenda[3,1], brenda)
}
