% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Onassis-methods.R
\docType{methods}
\name{scores}
\alias{scores}
\alias{scores<-}
\alias{scores<-,Onassis-method}
\alias{scores<-}
\alias{scores,Onassis-method}
\alias{scores}
\title{Method scores}
\usage{
scores(object)

scores(object) <- value

\S4method{scores}{Onassis}(object) <- value

\S4method{scores}{Onassis}(object)
}
\value{
The Onassis object

Ontology object
}
\description{
Method scores

Method scores<-

This method sets the score of Onassis

This method shows the score matrix in Onassis.
}
\examples{
onassis <- Onassis()
scores(onassis) <- matrix()

o <- Onassis()
scores(o)

}
