% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMoptions-methods.R
\docType{methods}
\name{SearchStrategy,CMoptions-method}
\alias{SearchStrategy,CMoptions-method}
\title{\code{SearchStrategy}}
\usage{
\S4method{SearchStrategy}{CMoptions}(x)
}
\arguments{
\item{x}{instance of class \code{\link{CMoptions}}}
}
\value{
The searchStrategy corresponding to the current options when used as get, the new CMoptions object with updated parameters when used to set.
}
\description{
Method to get and set the searchStrategy parameter
}
\examples{
opts <- CMoptions()
SearchStrategy(opts)

}
