% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/CMoptions-methods.R
\docType{methods}
\name{SearchStrategy}
\alias{SearchStrategy}
\alias{SearchStrategy<-}
\alias{SearchStrategy,CMoptions-method}
\alias{SearchStrategy<-,CMoptions-method}
\title{Method SearchStrategy}
\usage{
SearchStrategy(x)

SearchStrategy(x) <- value

\S4method{SearchStrategy}{CMoptions}(x)

\S4method{SearchStrategy}{CMoptions}(x) <- value
}
\arguments{
\item{x}{instance of class \code{\link{CMoptions}}}

\item{value}{The matching strategy for finding concepts in the input text
\itemize{
\item{CONTIGUOUS_MATCH}{Longets match of contiguous tokens within enclosing span}
\item{SKIP_ANY_MATCH}{Longest match of not-necessarily contiguous tokens}
\item{SKIP_ANY_MATCH_ALLOW_OVERLAP}{Longest match of not-necessarily contiguous tokens, overlapping matches are allowed}
}}
}
\value{
The SearchStrategy corresponding to the current options when used as get, the new CMoptions object with updated parameters when used to set.
}
\description{
Method to get and set the SearchStrategy parameter
}
\examples{
opts <- CMoptions()
SearchStrategy(opts)

opts <- CMoptions()
SearchStrategy(opts) <- 'SKIP_ANY_MATCH_ALLOW_OVERLAP'
}
