% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllFunctions.R
\name{showCMoptions}
\alias{showCMoptions}
\title{\code{showCMoptions}}
\usage{
showCMoptions()
}
\value{
list of possible options that one cas net to run the entity finder
}
\description{
This function shows the list of possible combinations of options to run the entity finder
}
\examples{
showCMoptions()
}
