% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/similarity-methods.R
\docType{methods}
\name{showOpts}
\alias{showOpts}
\alias{showOpts,Similarity-method}
\title{Method showOpts}
\usage{
showOpts(object)

\S4method{showOpts}{Similarity}(object)
}
\arguments{
\item{object}{Similarity instance}
}
\value{
the list of pairwise, information content and groupwise measures to compute the semantic similarities
}
\description{
Method showOpts

This method shows a list of the possible measures to compute pairwise and groupwise semantic similarity between concepts
}
\examples{
sim <- new('Similarity')
showOpts(sim)

}
