% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/similarity-methods.R
\docType{methods}
\name{sim}
\alias{sim}
\alias{sim,Similarity,character,character-method}
\alias{sim-method}
\title{Method sim}
\usage{
sim(object, term1, term2)

\S4method{sim}{Similarity,character,character}(object, term1, term2)
}
\arguments{
\item{object}{instance of the class \code{\link{Similarity-class}}}

\item{term1}{The URI of the ontology term in the format http://purl.obolibrary.org/obo/Ontology_id (e.g http://purl.obolibrary.org/obo/BTO_0004732)}

\item{term2}{The URI of the ontology term}
}
\value{
the semantic similarity of the two provided concepts
}
\description{
Method sim

This method computes the semantic similarity between two terms of a given ontology.
}
\examples{
sim <- new('Similarity')
dict_file <-  file.path(getwd(), 'BrendaTissue.obo')
 if(!file.exists(dict_file))
   download.file('https://sourceforge.net/projects/onassis/files/BrendaTissue.obo',
    destfile=file.path(getwd(), 'BrendaTissue.obo'))
ontology(sim) <- dict_file
pairwiseConfig(sim) <- showOpts(sim)$pairwiseMeasures[9]
}
