% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Onassis-methods.R
\docType{methods}
\name{simil}
\alias{simil}
\alias{simil<-}
\alias{simil<-,Onassis-method}
\alias{simil<-}
\alias{simil,Onassis-method}
\alias{simil}
\title{Method simil}
\usage{
simil(object)

simil(object) <- value

\S4method{simil}{Onassis}(object) <- value

\S4method{simil}{Onassis}(object)
}
\value{
The Onassis object

Ontology object
}
\description{
Method simil

Method simil<-

This method sets the similarity of Onassis

This method shows the similarity matrix in Onassis.
}
\examples{
onassis <- Onassis()
simil(onassis) <- matrix()

o <- Onassis()
simil(o)

}
