% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{Similarity-class}
\alias{Similarity-class}
\title{Similarity class to compute similarities between concepts in ontologies and samples annotated with different concepts
\code{Similarity-class}}
\description{
Similarity is a class that wraps some methods of the Java library slib \url{http://www.semantic-measures-library.org/sml/}. Starting from OBO ontologies it is possible to build semantic graphs that allow the computation of different similarity measures between concepts belonging to the same ontology, group of concepts, samples annotated with different ontology concepts. Further details about the graph based semantic similarity measures are available at \url{http://www.semantic-measures-library.org/sml/index.php?q=doc_graph_based_advanced}
}
\details{
The following methods can be applied to Similarity

\code{\link{ontology<-}} \cr
\code{\link{pairwiseConfig}} \cr
\code{\link{groupConfig}} \cr
\code{\link{sim}} \cr
\code{\link{groupsim}} \cr
\code{\link{samplesim}} \cr
To see the available similarity measures run the function \code{\link{listSimilarities}}
}
\section{Slots}{

\describe{
\item{\code{similarityInstance}}{The Java reference to the Java Similarity class.}

\item{\code{pairwiseConfig}}{The list of measures used to compute the semantic similarity between two concpets in the same ontology.}

\item{\code{pairwiseConfigRef}}{The reference to the Java object of type CMconf corresponding to the pairwise configuration}

\item{\code{groupConfig}}{The groupwise configuration to compute the semantic similarity between groups of concepts.}

\item{\code{icConfig}}{The information content measure}

\item{\code{groupwiseConfigRef}}{The reference to the Java configuration object for the computation of semantic similarity between groups of concepts}

\item{\code{ontology}}{The ontology to compute semantic similarities}
}}

\examples{
sim <- new('Similarity')

}
