% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllFunctions.R
\name{showSimilarities}
\alias{showSimilarities}
\title{\code{showSimilarities}}
\usage{
showSimilarities()
}
\value{
the list of options
}
\description{
This function shows the list of pairwise, information content and groupwise measures to compute the semantic similarities
}
\examples{
measures <- showSimilarities()

}
