% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllConstructors.R
\name{Similarity}
\alias{Similarity}
\title{\code{Similarity}}
\usage{
Similarity(ontology, termlist1, termlist2, annotatedtab = NA,
  icConf = "seco", pairConf = "lin", groupConf = "ui")
}
\arguments{
\item{ontology}{The ontology file to create the DAG to compute similarities}

\item{termlist1}{The single concept or vector of concepts in the first set or the name of a sample if annotatedtab contains an annotated table}

\item{termlist2}{The single concept or vector of concepts in the second set or the name of a sample if annotatedtab contains an annotated table}

\item{annotatedtab}{The table of annotation of samples or entities}

\item{icConf}{the information content configuration}

\item{pairConf}{the pairwise configuration}

\item{groupConf}{the groupwise configuration}
}
\value{
Measure of the similarity bewtween the concepts passed as input
}
\description{
this constructr initializes the Similarity class to compute the similarity between couple of terms, couple of samples, or group of terms
}
\examples{
obo <- system.file('extdata', 'sample.cs.obo', package='OnassisJavaLibs')
sample_dict <- CMdictionary(input=obo, outputDir=getwd(), synonymType='ALL')
myopts <- CMoptions(paramValueIndex=40)
term_list1 <- c('http://purl.obolibrary.org/obo/CL_0000000', 'http://purl.obolibrary.org/obo/CL_0000236')
term_list2 <- c('http://purl.obolibrary.org/obo/CL_0000542')
sim <- Similarity(obo, term_list1, term_list2)

}
