% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMoptions-methods.R
\docType{methods}
\name{Stemmer,CMoptions-method}
\alias{Stemmer,CMoptions-method}
\title{\code{Stemmer}}
\usage{
\S4method{Stemmer}{CMoptions}(x)
}
\arguments{
\item{x}{instance of class \code{\link{CMoptions}}}
}
\value{
The Stemmer corresponding to the current options when used as get, the new CMoptions object with updated parameters when used to set.
}
\description{
Method to get and set the Stemmer parameter
}
\examples{
opts <- CMoptions()
Stemmer(opts)

}
