% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/CMoptions-methods.R
\docType{methods}
\name{Stemmer}
\alias{Stemmer}
\alias{Stemmer<-}
\alias{Stemmer<-,CMoptions-method}
\title{Method Stemmer}
\usage{
Stemmer(x)

Stemmer(x) <- value

\S4method{Stemmer}{CMoptions}(x) <- value
}
\arguments{
\item{value}{\itemize{
\item BIOLEMMATIZER {A stemmer specific for biomedical literature}
\item PORTER {A stemmer that removes the commoner morphological and inflexional endings from words in English}
\item NONE {No word stemming}
}}
}
\description{
Method Stemmer

Method Stemmer<-

\code{Stemmer<-}
}
\examples{
opts <- CMoptions()
Stemmer(opts) <- 'PORTER'
}
