% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMoptions-methods.R
\docType{methods}
\name{StopWords,CMoptions-method}
\alias{StopWords,CMoptions-method}
\title{\code{StopWords}}
\usage{
\S4method{StopWords}{CMoptions}(x)
}
\arguments{
\item{x}{instance of class \code{\link{CMoptions}}}
}
\value{
The StopWords corresponding to the current options when used as get, the new CMoptions object with updated parameters when used to set.
}
\description{
Method to get and set the StopWords parameter
}
\examples{
opts <- CMoptions()
StopWords(opts)

}
