% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/CMoptions-methods.R
\docType{methods}
\name{StopWords}
\alias{StopWords}
\alias{StopWords<-}
\alias{StopWords<-,CMoptions-method}
\title{Method StopWords}
\usage{
StopWords(x)

StopWords(x) <- value

\S4method{StopWords}{CMoptions}(x) <- value
}
\arguments{
\item{value}{\itemize{
\item PUBMED {A list of stop words obtained analyzing Pubmed papers}
\item NONE {No stop words }
}}
}
\description{
Method StopWords

Method StopWords<-

\code{StopWords<-}
}
\examples{
opts <- CMoptions()
StopWords(opts) <- 'NONE'
}
