% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMoptions-methods.R
\docType{methods}
\name{SynonymType,CMoptions-method}
\alias{SynonymType,CMoptions-method}
\title{\code{SynonymType}}
\usage{
\S4method{SynonymType}{CMoptions}(x)
}
\arguments{
\item{x}{instance of class \code{\link{CMoptions}}}
}
\value{
The SynonymType corresponding to the current options when used as get, the new CMoptions object with updated parameters when used to set.
}
\description{
Method to get and set the SynonymType parameter
}
\examples{
opts <- CMoptions()
SynonymType(opts)

}
