% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/EntityFinder-methods.R
\docType{methods}
\name{typeSystemRef}
\alias{typeSystemRef}
\alias{typeSystemRef<-}
\alias{typeSystemRef,EntityFinder-method}
\alias{typeSystemRef<-,EntityFinder-method}
\title{Method typeSystemRef}
\usage{
typeSystemRef(x)

typeSystemRef(x) <- value

\S4method{typeSystemRef}{EntityFinder}(x)

\S4method{typeSystemRef}{EntityFinder}(x) <- value
}
\arguments{
\item{x}{instance of the class EntityFinder}

\item{value}{the java type system to detect concepts from ontologies.}
}
\value{
the reference to the Java type system currently set

The updated EntityFinder S4 object
}
\description{
This method sets the type system to the ccp-nlp one to run the EntityFinder

This method sets the type system to the ccp-nlp one to run the EntityFinder
}
\examples{
ef <- new('EntityFinder')
typeSystemRef(ef)
ef <- new('EntityFinder')
type_system_array_list <- .jnew('java/util/ArrayList')
ccp_nlp_type_system <- .jfield('edu/ucdenver/ccp/nlp/uima/util/TypeSystemUtil',
 name = 'CCP_TYPE_SYSTEM')
sentence_detector_type_system_str <- 'org.cleartk.token.type.Sentence'
conceptmapper_type_system <-
'edu.ucdenver.ccp.nlp.wrapper.conceptmapper.TypeSystem'
dictTerm <- 'analysis_engine.primitive.DictTerm'
tokenizer <- 'org.apache.uima.conceptMapper.support.tokenizer.TokenAnnotation'
vector_of_ts <- c(ccp_nlp_type_system, sentence_detector_type_system_str,
 conceptmapper_type_system, dictTerm, tokenizer)
type_system_description <-
J('org/uimafit/factory/TypeSystemDescriptionFactory')$createTypeSystemDescription(vector_of_ts)
typeSystemRef(ef) <- type_system_description
}
