% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Onassis-methods.R
\docType{methods}
\name{annot}
\alias{annot}
\alias{annot,data.frame,character,character-method}
\alias{annot}
\title{\code{annot}}
\usage{
annot(input = NA, dictType = NA, dictionary = NA, ...)

\S4method{annot}{data.frame,character,character}(input, dictType = "OBO",
  dictionary = NA, dictoutdir = getwd(), d_synonymtype = "EXACT",
  taxID = 0, annot_out = getwd(), paramValueIndex = 0,
  searchstrategy = "CONTIGUOUS_MATCH", casematch = "CASE_INSENSITIVE",
  stemmer = "PORTER", stopwords = "NONE", orderindependentlookup = "OFF",
  findallmatches = "NO", e_synonymtype = "EXACT_ONLY",
  multipledocs = FALSE, disease = FALSE)
}
\arguments{
\item{input}{A data frame where the first column is the ID of the sample or document to annotate}

\item{dictType}{the type of input dictionary
\describe{
\item{OBO}{A dictionary that has been created by An OBO file}
\item{ENTREZ}{Entrez genes dictionary}
\item{TARGET}{Entrez genes dictionary, Histone marks and Histone modifications}
\item{CMDICT}{A previously created dictionary file in the Conceptmapper XML format}
}}

\item{dictionary}{The local OBO/OWL ontology to be converted into an XML Conceptmapper dictionary or the URL to download the file. If NA is passed and the \code{dicType} parameter is not the generic OBO then the method tries to download the corresponding dictionary from the available repositories. For ENTREZ and TARGET dictionary types a file named gene_info.gz is automatically downloaded from \url{ftp://ncbi.nlm.nih.gov/gene/data/gene_info.gz} if not provided by the user. Alternatively an annotation package of the type Org.xx.eg.db from Bioconductor can be used. In this case the gene ids and their alternative identifiers will be retrieved from the annotation database without the need of downloading a gene_info file.}

\item{dictoutdir}{Optional parameter to specify the location where the Conceptmapper dictionary file will be stored. Defaults to current working directory.}

\item{d_synonymtype}{The type of synonyms to consider when building the dictionary for Conceptmapper. For further detail \url{http://owlcollab.github.io/oboformat/doc/obo-syntax.html}. Default: EXACT
\describe{
\item{EXACT}{}
\item{ALL}{}
}}

\item{taxID}{the taxonomy identifier of the organism when the dictionary type is ENTREZ or TARGET. If 0 all the taxonomies will be included in the new dictionary.}

\item{disease}{A logical value set to TRUE if the annotation requires the 'Healthy' condition to be found.}
}
\description{
This method returns and sets the location of the dictionary.
}
\examples{
geo_chip <- readRDS(system.file('extdata', 'vignette_data',
'GEO_human_chip.rds', package='Onassis'))

obo <- system.file('extdata', 'sample.cs.obo', package='OnassisJavaLibs')
onassis_results <- annot(geo_chip, "OBO", dictionary=obo)
entities <- entities(onassis_results)
entities <- entities[sample(nrow(entities), 30),]
}
