% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllFunctions.R
\name{annotate}
\alias{annotate}
\title{\code{annotate}}
\usage{
annotate(inputFileorDf, dictionary, options = NA, outDir = tempdir(),
  multipleDocs = FALSE)
}
\arguments{
\item{inputFileorDf}{the file, directory, or data frame to annotate}

\item{dictionary}{the Conceptmapper dictionary file}

\item{options}{the options to run the finder}

\item{outDir}{the directory where to store the output files from onassis}

\item{multipleDocs}{TRUE if the input file is one but contains multiple documents}
}
\value{
dataframe of annotations
}
\description{
this function runs the entityfinder on given input
}
\examples{
dict_file <-  file.path(getwd(), 'cmDict-BrendaTissue.xml')
 if(!file.exists(dict_file))
   download.file('https://sourceforge.net/projects/onassis/files/cmDict-BrendaTissue.xml',
    destfile=file.path(getwd(), 'cmDict-BrendaTissue.xml'))
brenda_dict <- CMdictionary(inputFile=file.path(getwd(),
'cmDict-BrendaTissue.xml'), dictType = 'CMDICT')
myopts <- CMoptions(40)
tissue_entities <- annotate(system.file('extdata',
'testsamples', package='Onassis'), options = myopts,
 dictionary=brenda_dict)

}
