% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/CMoptions-methods.R
\docType{class}
\name{CMoptions-class}
\alias{CMoptions-class}
\alias{show,CMoptions-method}
\title{Class to set the options to run the EntityFinder}
\usage{
\S4method{show}{CMoptions}(object)
}
\arguments{
\item{object}{CMoptions instance}
}
\value{
the list of options
}
\description{
CMoptions is a class that represents Conceptmapper configurations. It allows users to set the possible combinations of different parameters for Conceptmapper running.

This method shows the list of options to run the Entity finder
}
\details{
The following methods can be applied to CMoptions

\code{\link{show}}  \cr
\code{\link{paramValueIndex}} \cr
\code{\link{paramValueIndex<-}} \cr
\code{\link{SearchStrategy}} \cr
\code{\link{SearchStrategy<-}} \cr
\code{\link{CaseMatch}} \cr
\code{\link{CaseMatch<-}} \cr
\code{\link{Stemmer}} \cr
\code{\link{Stemmer<-}} \cr
\code{\link{StopWords}} \cr
\code{\link{StopWords<-}} \cr
\code{\link{OrderIndependentLookup}} \cr
\code{\link{OrderIndependentLookup<-}} \cr
\code{\link{FindAllMatches}} \cr
\code{\link{FindAllMatches<-}} \cr
\code{\link{SynonymType}} \cr
\code{\link{SynonymType<-}} \cr
}
\section{Slots}{

\describe{
\item{\code{paramValueIndex}}{An integer value to index the 576 parameter combinations}

\item{\code{SearchStrategy}}{The matching strategy for finding concepts in the input text
\itemize{
\item{CONTIGUOUS_MATCH}{Longets match of contiguous tokens within enclosing span}
\item{SKIP_ANY_MATCH}{Longest match of not-necessarily contiguous tokens}
\item{SKIP_ANY_MATCH_ALLOW_OVERLAP}{Longest match of not-necessarily contiguous tokens, overlapping matches are allowed}
}}

\item{\code{CaseMatch}}{\itemize{
\item{CASE_IGNORE}{Fold everything to lowercase for matching}
\item{CASE_INSENSITIVE}{Fold only tokens with initial caps to lowercase}
\item{CASE_FOLD_DIGITS}{Fold all (and only) tokens with a digit}
\item{CASE_SENSITIVE}{Perform no case folding}
}}

\item{\code{Stemmer}}{\itemize{
\item BIOLEMMATIZER {A stemmer specific for biomedical literature}
\item PORTER {A stemmer that removes the commoner morphological and inflexional endings from words in English}
\item NONE {No word stemming}
}}

\item{\code{StopWords}}{\itemize{
\item PUBMED {A list of stop words obtained analyzing Pubmed papers}
\item NONE {No stop words }
}}

\item{\code{OrderIndependentLookup}}{\itemize{
\item ON {Ordering within span is ignored (i.e. 'Breast cancer' would equal 'Cancer breast') }
\item OFF {Ordering is taken into consideration}
}}

\item{\code{FindAllMatches}}{\itemize{
\item YES {All the matches within the span are found }
\item NO {Only the longest match within the span will be returned}
}}

\item{\code{SynonymType}}{\itemize{
\item EXACT_ONLY {Only exact synonyms are considered }
\item ALL {All synonym types are included}
}}
}}

\examples{
options <- new('CMoptions')
opt <- CMoptions()
show(opt)
}
