% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Onassis-methods.R
\docType{methods}
\name{compare}
\alias{compare}
\alias{compare}
\alias{mergeonassis}
\title{\code{compare}}
\usage{
compare(onassis, ...)

\S4method{compare}{Onassis}(onassis, score_matrix = NA, by = "row",
  fun_name = "wilcox.test")
}
\description{
This method compares a list of scored entities
}
\examples{
geo_chip <- readRDS(system.file('extdata', 'vignette_data','GEO_human_chip.rds', package='Onassis'))
geo_chip <- geo_chip[sample(nrow(geo_chip), 30) ,]
obo1 <- system.file('extdata', 'sample.cs.obo', package='OnassisJavaLibs')
obo2 <- system.file('extdata', 'sample.do.obo', package='OnassisJavaLibs')
onassis_results1 <- annot(geo_chip, "OBO", dictionary=obo1)
onassis_results2 <- annot(geo_chip, "OBO", dictionary=obo2)
n <- length(unique(geo_chip$sample_accession))
m <- 50
score_matrix <-   matrix(sample(0:1, m * n, replace = TRUE), m, n)
colnames(score_matrix) <- unique(geo_chip$sample_accession)
rownames(score_matrix) <- paste0('gene_', seq(1, m, 1))
my_onassis <- mergeonassis(onassis1 = onassis_results1, onassis2 = onassis_results2)
scores(my_onassis) <- score_matrix

comparisons <- compare(my_onassis)
comparisons2 <- compare(my_onassis, by='col', fun_name='kruskal.test')
scores(onassis_results1) <- score_matrix
comparisons3 <- compare(onassis_results1)
comparisons4 <- compare(onassis_results1, by='col', fun_name='kruskal.test')
}
