% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllFunctions.R
\name{dictionaryInfo}
\alias{dictionaryInfo}
\title{\code{dictionaryInfo}}
\usage{
dictionaryInfo(CMdictionary)
}
\arguments{
\item{CMdictionary}{CMdictionary object of the conceptmapper dictionary}
}
\value{
list of details about the dictionary
}
\description{
This function shows the list of details of the conceptmapper dictionary
}
\examples{
dict_file <-  file.path(getwd(), 'BrendaTissue.obo')
 if(!file.exists(dict_file))
   download.file('https://sourceforge.net/projects/onassis/files/BrendaTissue.obo',
    destfile=file.path(getwd(), 'BrendaTissue.obo'))
ontology_file <- file.path(getwd(), 'BrendaTissue.obo')
dict <- CMdictionary(inputFile=ontology_file, outputdir=getwd(), synonymType='ALL')
dictionaryInfo(dict)
}
