% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Onassis-methods.R
\docType{methods}
\name{entities}
\alias{entities}
\alias{entities<-}
\alias{entities<-,Onassis-method}
\alias{entities,Onassis-method}
\title{\code{entities<-}}
\usage{
entities(object)

entities(object) <- value

\S4method{entities}{Onassis}(object) <- value

\S4method{entities}{Onassis}(object)
}
\arguments{
\item{object}{instance of class \code{\link{Onassis-class}}}

\item{value}{a data frame with annotated entities}
}
\value{
the entities of the Onassis object in case of get and the Onassis object with new entities in case of set
}
\description{
Method to get and set the \code{entities} slot of the class \code{\link{Onassis-class}}
}
\examples{
onassis <- Onassis()
entities(onassis) <- data.frame()

o <- Onassis()
entities(o)

}
