% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Onassis-methods.R
\docType{methods}
\name{filterconcepts}
\alias{filterconcepts}
\alias{filterconcepts}
\title{\code{filterconcepts}}
\usage{
filterconcepts(onassis, concepts_to_filter)

\S4method{filterconcepts}{Onassis}(onassis, concepts_to_filter = c())
}
\arguments{
\item{onassis}{An object of class Onassis with already annotated entities}

\item{concepts_to_filter}{A vector with unwanted concepts}
}
\description{
This method filter unwanted concepts from the entities of an onassis object
}
\examples{
geo_chip <- readRDS(system.file('extdata', 'vignette_data',
'GEO_human_chip.rds', package='Onassis'))
geo_chip <- geo_chip[sample(nrow(geo_chip), 15) ,]
obo <- system.file('extdata', 'sample.cs.obo', package='OnassisJavaLibs')
onassis_results <- annot(geo_chip, "OBO", dictionary=obo)
filtered_onassis <- filterconcepts(onassi_results, c('cell'))

}
