% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/CMoptions-methods.R
\docType{methods}
\name{FindAllMatches}
\alias{FindAllMatches}
\alias{FindAllMatches<-}
\alias{FindAllMatches,CMoptions-method}
\alias{FindAllMatches<-,CMoptions-method}
\title{Method FindAllMatches}
\usage{
FindAllMatches(x)

FindAllMatches(x) <- value

\S4method{FindAllMatches}{CMoptions}(x)

\S4method{FindAllMatches}{CMoptions}(x) <- value
}
\arguments{
\item{x}{instance of class \code{\link{CMoptions}}}

\item{value}{\itemize{
\item YES {All the matches within the span are found }
\item NO {Only the longest match within the span will be returned}
}}
}
\value{
The FindAllMatches corresponding to the current options when used as get, the new CMoptions object with updated parameters when used to set.
}
\description{
Method to get and set the FindAllMatches parameter
}
\examples{
opts <- CMoptions()
FindAllMatches(opts)
opts <- CMoptions()
FindAllMatches(opts) <- 'YES'
}
