% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/similarity-methods.R
\docType{methods}
\name{groupsim}
\alias{groupsim}
\alias{groupsim,Similarity,character,character-method}
\title{Method groupsim}
\usage{
groupsim(object, termList1, termList2)

\S4method{groupsim}{Similarity,character,character}(object, termList1,
  termList2)
}
\arguments{
\item{object}{instance of class \code{\link{Similarity-class}}}

\item{termList1}{A vector of URIs of ontology terms in the format \url{http://purl.obolibrary.org/obo/Ontology_id} (e.g \url{http://purl.obolibrary.org/obo/BTO_0004732})}

\item{termList2}{A vector of URIs of ontology terms}
}
\value{
the semantic similarity of the two provided groups of concepts
}
\description{
This method computes the semantic similarity between two groups of terms of a given ontology.
}
\examples{

sim <- new('Similarity')
obo <- system.file('extdata', 'sample.cs.obo', package='OnassisJavaLibs')
ontology(sim) <- obo
pairwiseConfig(sim) <- listSimilarities()$pairwiseMeasures[9]
groupConfig(sim) <- listSimilarities()$groupwiseMeasures[3]
similarity <- groupsim(sim, c('http://purl.obolibrary.org/obo/CL_0000542',
 'http://purl.obolibrary.org/obo/CL_0000236'),
c('http://purl.obolibrary.org/obo/CL_0000000'))
similarity

}
