% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{package}
\name{OnASSiS}
\alias{OnASSiS}
\alias{Onassis-package}
\alias{OnASSiS-package}
\title{OnASSiS (Ontology Annotations and Semantic Similarity software)}
\description{
OnASSiS (Ontology Annotations and Semantic Similarity software) is a package for the annotation of any given text with concepts from biomedical ontologies that also provides features to relate the concepts using semantic similarity metrics.
}
\details{
OnASSiS package


OnASSiS (Ontology Annotations and Semantic Similarity software) is a package that uses Conceptmapper, an Apache UIMA (Unstructured Information Management Architecture) \url{https://uima.apache.org/downloads/sandbox/ConceptMapperAnnotatorUserGuide/ConceptMapperAnnotatorUserGuide.html} dictionary lookup tool to retrieve dictionary terms in a given text.

In particular a Conceptmapper wrapper specific for the biomedical domain, ccp-nlp, (\url{https://github.com/UCDenver-ccp/ccp-nlp}) has been personalized to retrieve concepts from OBO ontologies in a given text with different options.

The package also provides the possibility to annotate Gene Expression Omnibus (GEO) metadata for stored experiments and samples.

Different annotated sets of text can be then compared using semantic similarity metrics based on the structure of the biomedical ontologies.
The semantic similarity module has been obtained using the Java slib  (\url{http://www.semantic-measures-library.org/sml/})
}
