% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/similarity-methods.R
\docType{methods}
\name{ontology<-}
\alias{ontology<-}
\alias{ontology<-,Similarity-method}
\title{Method ontology<-}
\usage{
ontology(object) <- value

\S4method{ontology}{Similarity}(object) <- value
}
\arguments{
\item{object}{Instance of the class \code{\link{Similarity-class}}}

\item{value}{the ontology file}
}
\value{
The Similarity object where 'graph' slot refers to the Java graph created
}
\description{
Method ontology<-

This method creates a semantic graph to compute semantic similarity between concepts. It takes as input an OBO ontology in RDF, OWL or OBO format.
}
\examples{
sim <- new('Similarity')
dict_file <-  file.path(getwd(), 'BrendaTissue.obo')
 if(!file.exists(dict_file))
   download.file('https://sourceforge.net/projects/onassis/files/BrendaTissue.obo',
    destfile=file.path(getwd(), 'BrendaTissue.obo'))
ontology(sim) <- dict_file

}
