% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/CMoptions-methods.R
\docType{methods}
\name{OrderIndependentLookup}
\alias{OrderIndependentLookup}
\alias{OrderIndependentLookup<-}
\alias{OrderIndependentLookup<-,CMoptions-method}
\title{Method OrderIndependentLookup}
\usage{
OrderIndependentLookup(x)

OrderIndependentLookup(x) <- value

\S4method{OrderIndependentLookup}{CMoptions}(x) <- value
}
\arguments{
\item{value}{\itemize{
\item ON {Ordering within span is ignored (i.e. 'Breast cancer' would equal 'Cancer breast') }
\item OFF {Ordering is taken into consideration}
}}
}
\description{
Method OrderIndependentLookup

Method OrderIndependentLookup<-

\code{OrderIndependentLookup<-}
}
\examples{
opts <- CMoptions()
OrderIndependentLookup(opts) <- 'YES'
}
