% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/CMoptions-methods.R
\docType{methods}
\name{paramValueIndex}
\alias{paramValueIndex}
\alias{paramValueIndex<-}
\alias{paramValueIndex,CMoptions-method}
\alias{paramValueIndex<-,CMoptions-method}
\title{Method paramValueIndex}
\usage{
paramValueIndex(x)

paramValueIndex(x) <- value

\S4method{paramValueIndex}{CMoptions}(x)

\S4method{paramValueIndex}{CMoptions}(x) <- value
}
\arguments{
\item{x}{instance of class \code{\link{CMoptions}}}

\item{value}{Index corresponding to a given parameter combination}
}
\value{
The paramValueIndex corresponding to the current options when used as get, the new CMoptions object with updated parameters when used to set.
}
\description{
Method to get and set the parameter combination index corresponding to a given parameter combination. The value of the paramValueIndex lays in the range [0:575]
}
\examples{
opts <- CMoptions()
paramValueIndex(opts)

opts <- CMoptions()
paramValueIndex(opts) <- 2
}
