% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Onassis-methods.R
\docType{methods}
\name{scores}
\alias{scores}
\alias{scores<-}
\alias{scores<-,Onassis-method}
\alias{scores,Onassis-method}
\title{\code{scores<-}}
\usage{
scores(object)

scores(object) <- value

\S4method{scores}{Onassis}(object) <- value

\S4method{scores}{Onassis}(object)
}
\arguments{
\item{object}{instance of class \code{\link{Onassis-class}}}

\item{value}{a matrix of scores}
}
\value{
the matrix of scores in case of get and the new \code{\link{Onassis-class}} object in case of set
}
\description{
This method gets and sets the \code{scores} slot of a class \code{\link{Onassis-class}}

\code{scores}
}
\examples{
onassis <- Onassis()
scores(onassis) <- matrix()

o <- Onassis()
scores(o)

}
