% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/CMoptions-methods.R
\docType{methods}
\name{SearchStrategy}
\alias{SearchStrategy}
\alias{SearchStrategy<-}
\alias{SearchStrategy<-,CMoptions-method}
\title{Method SearchStrategy}
\usage{
SearchStrategy(x)

SearchStrategy(x) <- value

\S4method{SearchStrategy}{CMoptions}(x) <- value
}
\arguments{
\item{value}{The matching strategy for finding concepts in the input text
\itemize{
\item{CONTIGUOUS_MATCH}{Longets match of contiguous tokens within enclosing span}
\item{SKIP_ANY_MATCH}{Longest match of not-necessarily contiguous tokens}
\item{SKIP_ANY_MATCH_ALLOW_OVERLAP}{Longest match of not-necessarily contiguous tokens, overlapping matches are allowed}
}}
}
\description{
Method SearchStrategy

Method SearchStrategy<-

\code{SearchStrategy<-}
}
\examples{
opts <- CMoptions()
SearchStrategy(opts) <- 'SKIP_ANY_MATCH_ALLOW_OVERLAP'
}
