% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Onassis-methods.R
\docType{methods}
\name{sim}
\alias{sim}
\alias{sim,Onassis-method}
\alias{sim}
\title{\code{sim}}
\usage{
sim(onassis = NA, ...)

\S4method{sim}{Onassis}(onassis, iconf = "sanchez", pairconf = "lin",
  groupconf = "bma")
}
\description{
This method computes the similarities of the entities in an Onassis object.
}
\examples{
geo_chip <- readRDS(system.file('extdata', 'vignette_data', 'GEO_human_chip.rds', package='Onassis'))
obo <- system.file('extdata', 'sample.cs.obo', package='OnassisJavaLibs')
onassis_results <- annot(geo_chip, "OBO", dictionary=obo)
entities <- entities(onassis_results)
entities(onassis_results) <- entities[sample(nrow(entities), 30),]
onassis_results <- sim(onassis_results)

}
