% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/AllFunctions.R
\docType{class}
\name{Similarity-class}
\alias{Similarity-class}
\alias{similarity}
\title{Similarity class to compute similarities between concepts in ontologies and samples annotated with different concepts
\code{Similarity-class}}
\usage{
similarity(ontologyFile, termlist1, termlist2, annotatedtab = NA,
  pairConf = c("resnik", "seco"), groupConf = "ui")
}
\arguments{
\item{ontologyFile}{the file in OBO or RDF format to build the graph}

\item{termlist1}{The URI or character vector of URIs belonging to the first set or the sample id of an annotated data frame}

\item{termlist2}{The URI or character vector of URIs belonging to the second  set or the sample id of the annotated data frame}

\item{annotatedtab}{The data frame of}

\item{pairConf}{a configuration for the pairwise meausres. Defaults to resnik and seco.}

\item{groupConf}{one of the allowed configurations for groupwise measures}
}
\value{
the similarity value based on the configured measures
}
\description{
Similarity is a class that wraps some methods of the Java library slib \url{http://www.semantic-measures-library.org/sml/}. Starting from OBO ontologies it is possible to build semantic graphs that allow the computation of different similarity measures between concepts belonging to the same ontology, group of concepts, samples annotated with different ontology concepts. Further details about the graph based semantic similarity measures are available at \url{http://www.semantic-measures-library.org/sml/index.php?q=doc_graph_based_advanced}

this function computes the similarity between couple of terms, couple of samples, or group of terms
}
\details{
The following methods can be applied to Similarity

\code{\link{ontology<-}} \cr
\code{\link{pairwiseConfig}} \cr
\code{\link{groupwiseConfig}} \cr
\code{\link{showOpts}} \cr
\code{\link{sim}} \cr
\code{\link{groupsim}} \cr
\code{\link{samplesim}} \cr
}
\section{Slots}{

\describe{
\item{\code{similarityInstance}}{The Java reference to the Java Similarity class.}

\item{\code{pairwiseConfig}}{The list of measures used to compute the semantic similarity between two concpets in the same ontology.}

\item{\code{pairwiseConfigRef}}{The reference to the Java object of type CMconf corresponding to the pairwise configuration}

\item{\code{groupConfig}}{The groupwise configuration to compute the semantic similarity between groups of concepts.}

\item{\code{icConfig}}{The information content measure}

\item{\code{graph}}{The reference to the Java object corresponding to the DAG created from the ontology.}

\item{\code{groupwiseConfigRef}}{The reference to the Java configuration object for the computation of semantic similarity between groups of concepts}
}}

\examples{
sim <- new('Similarity')

dict_file <-  file.path(getwd(), 'BrendaTissue.obo')
 if(!file.exists(dict_file))
   download.file('https://sourceforge.net/projects/onassis/files/BrendaTissue.obo',
    destfile=file.path(getwd(), 'BrendaTissue.obo'))
ontology_file <- file.path(getwd(), 'BrendaTissue.obo')
term_list1 <- c('http://purl.obolibrary.org/obo/BTO_0001546',
'http://purl.obolibrary.org/obo/BTO_0000664')
term_list2 <- c('http://purl.obolibrary.org/obo/BTO_0000759',
'http://purl.obolibrary.org/obo/BTO_0000759')
sim <- similarity(ontology_file, term_list1, term_list2)
}
