% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/CMoptions-methods.R
\docType{methods}
\name{SynonymType}
\alias{SynonymType}
\alias{SynonymType<-}
\alias{SynonymType<-,CMoptions-method}
\title{Method SynonymType}
\usage{
SynonymType(x)

SynonymType(x) <- value

\S4method{SynonymType}{CMoptions}(x) <- value
}
\arguments{
\item{value}{\itemize{
\item EXACT_ONLY {Only exact synonyms are considered }
\item ALL {All synonym types are included}
}}
}
\description{
Method SynonymType

Method SynonymType<-

\code{SynonymType<-}
}
\examples{
opts <- CMoptions()
SynonymType(opts) <- 'ALL'
}
